branch-siham-2: Correction error message delete file of reservations

+++ class/Controllers/ReservationsController.php
@@ -161,29 +161,35 @@ class ReservationsController
     {
         $html = '';
 
-        // If the form have been submitted and not empty :
-        if (!empty($_POST['id'])) {
-            // Clean and validate the inputs
-            $id = trim(htmlspecialchars(strip_tags($_POST['id'])));
+        // Check if the form has been submitted
+        if ($_SERVER["REQUEST_METHOD"] == "POST"){
+            // If fields have been submitted and not empty :
+            if (!empty($_POST['id'])) {
+                // Clean and validate the inputs
+                $id = trim(htmlspecialchars(strip_tags($_POST['id'])));
 
-            // Check if 'id' is a numeric value
-            if (is_numeric($id)) {
-                // Check if 'id' is a positive number
-                if ($id >= 0) {
-                    // Delete the reservation :
-                    $reservationsService = new ReservationsService();
-                    $isOk = $reservationsService->deleteReservation($id);
-                    if ($isOk) {
-                        $html = 'Réservation supprimé avec succès.';
+                // Check if 'id' is a numeric value
+                if (is_numeric($id)) {
+                    // Check if 'id' is a positive number
+                    if ($id >= 0) {
+                        // Delete the reservation :
+                        $reservationsService = new ReservationsService();
+                        $isOk = $reservationsService->deleteReservation($id);
+                        if ($isOk) {
+                            $html = 'Réservation supprimé avec succès.';
+                        } else {
+                            $html = 'Erreur lors de la suppression de la réservation.';
+                        }
                     } else {
-                        $html = 'Erreur lors de la suppression de la réservation.';
+                        $html = 'Erreur : L\'id doit être un nombre positif.';
                     }
                 } else {
-                    $html = 'Erreur : L\'id doit être un nombre positif.';
+                    $html = 'Erreur : L\'id doit être une valeur numérique.';
                 }
             } else {
-                $html = 'Erreur : L\'id doit être une valeur numérique.';
+                $html = 'Erreur : Aucun identifiant saisi';
             }
+
         }
         return $html;
     }
